+++ README.md
@@ -5,8 +5,8 @@
 - [x] 컴포넌트 단위로 뷰 및 테스트 코드 단순 분리
 - [x] SignUpTitle 테스트 코드 작성
 - [x] SignUpHelperText 테스트 코드 작성
-- [] 컴포넌트 공통 요소 추출
-- [] 컴포넌트 고유 속성 주입
+- [x] PasswordConfirmTextField 컴포넌트 고유 속성 주입
+- [] PasswordConfirmTextField 테스트 코드 작성
 - [] 테스트 코드 리팩터링
 - [] 유효성 검사 로직 테스트 코드 추가
 - [] SignUpButton 활성화 로직 추가
@@ -20,7 +20,11 @@
   컴포즈의 문제가 아니라 저의 명백한 실수네요.
 - 하나의 TextField로 모든 컴포넌트를 커버하기 위해서 불필요한 로직(복잡성!) 고민으로 보낸 시간이 아깝네요. 그 시간에 컴포넌트 단위로 나누는 연습에 더 집중했어야
   했는데...
--
+- 처음에는 비밀번호와 비밀번호 확인 헬퍼 텍스트에 동일한 기준으로 표시하고 비밀번호 확인에는 추가적으로 일치 여부 메시지를 표시하려고 했습니다. 그래서 비밀번호 확인 아래에
+  추가적인 헬퍼 텍스트필드가 있었습니다. 그렇다 보니 로직이 좀 더 복잡해졌습니다. 그런데 생각해 보니 이건 기획 기준에 따라 다르겠지만 굳이 비밀번호 확인에서는 영문과 숫자,
+  길이 제한에 대한 헬퍼 텍스트를 노출할 필요가 없다고 생각됩니다. 일치하지 않는다는 것은 그것을 이미 어기고 있다는 것이니까요. 그래서 그부분은 제거하고 일치 여부에 대한 헬퍼
+  텍스트 노출만 하는 방식으로 바꿔봤습니다.
+- UsernameTextField와 EmailTextField는 다시 합치고 싶지만 일단 분리했습니다.
 
 ## Step3 구현 목록
