+++ app/src/androidTest/java/nextstep/signup/view/EmailTextFieldTest.kt
@@ -1,9 +1,8 @@
-package nextstep.signup
+package nextstep.signup.view
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
-import nextstep.signup.view.EmailTextField
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
