+++ app/src/androidTest/java/nextstep/signup/view/PasswordTextFieldTest.kt
@@ -1,10 +1,9 @@
-package nextstep.signup
+package nextstep.signup.view
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextClearance
 import androidx.compose.ui.test.performTextInput
-import nextstep.signup.view.PasswordTextField
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
