+++ app/src/androidTest/java/nextstep/signup/view/SignUpHelperTextTest.kt
@@ -1,9 +1,8 @@
-package nextstep.signup
+package nextstep.signup.view
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.signup.view.SignUpHelperText
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
