+++ app/src/androidTest/java/nextstep/signup/view/SignUpTitleTest.kt
@@ -1,8 +1,7 @@
-package nextstep.signup
+package nextstep.signup.view
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.signup.view.SignUpTitle
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
