+++ app/src/main/java/nextstep/signup/TextFieldType.kt
@@ -4,5 +4,6 @@ sealed class TextFieldType {
     data object Username : TextFieldType()
     data object Email : TextFieldType()
     data object Password : TextFieldType()
+    data object PasswordConfirm : TextFieldType()
     data object Default : TextFieldType()
 }
